+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/CompaniesHouse/SubDtos/LimitedPartnershipsDto.cs
@@ -2,7 +2,7 @@
 {
     public class LimitedPartnershipsDto
     {
-        public int NumGenPartners { get; set; }
-        public int NumLimPartners { get; set; }
+        public decimal NumGenPartners { get; set; }
+        public decimal NumLimPartners { get; set; }
     }
 }
